pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "cython"]
build-backend = "setuptools.build_meta"

[project]
name = "wizardhtml"
version = "0.1.0"
description = "WHATWG-compliant HTML5 toolkit: DFA tokenizer, spec-guided tree builder, DOM, configurable serializer, high-level cleaner, pretty-printer, and HTML to Markdown."
license = { file = "LICENSE" }

readme = { file = "README.md", content-type = "text/markdown" }

authors = [{ name = "Mattia Rubino", email = "textwizard.dev@gmail.com" }]
requires-python = ">=3.9"
keywords = [
  "html",
  "html5",
  "whatwg",
  "html-parser",
  "tokenizer",
  "dfa",
  "dom",
  "tree-builder",
  "serializer",
  "pretty-printer",
  "pretty-printing",
  "formatter",
  "cleaner",
  "sanitizer",
  "html-cleaning",
  "html-to-markdown",
  "markdown-converter",
  "unicode",
  "html-entities",
  "rcdata",
  "svg",
  "mathml",
  "fragment-parsing",
  "dom-serialization"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Intended Audience :: Developers",
  "Topic :: Text Processing",
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"
]

dependencies = [
  "marisa-trie>=1.3.0",
  "pyahocorasick>=2.2.0", 
]

[project.urls]
Homepage = "https://github.com/textwizard-dev/wizardhtml"
Source = "https://github.com/textwizard-dev/wizardhtml"
Documentation = "https://wizardhtml.readthedocs.io/en/latest/"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"] 

[tool.setuptools.packages.find]
include = ["wizardhtml", "wizardhtml.*"]


[[tool.setuptools.ext-modules]]
name = "wizardhtml.utils.tw_html_parser.entities_wrapper"
sources = [
  "wizardhtml/utils/tw_html_parser/entities_wrapper.pyx",
  "wizardhtml/utils/tw_html_parser/entities_hash.c"
]
include-dirs = ["wizardhtml/utils/tw_html_parser"]